- Creada la primera version del paso a jvm (sin testear y a medias)    - Comentadas todas las instrucciones con lo que deben hacer. La instruccion suma puede servir de base para hacer el resto, pues es la unica de las complejas que esta implementada.    - He cambiado tambien los toChar, toNat, toInt y toFloat de los tipos de dato para adaptarlo a los castings que dice el enunciado de la practica.    - Haciendo esto anterior me di cuenta que no tenemos ninguna instruccion CastNat, que nos hace farta    - Al implementar las instrucciones intentad no usar el casting de los datos, aplicad las operaciones solo si los dos operando son del mismo tipo (como se hace en la suma). En el foro dice el tomatoma que es mejor hacerlo asi (cuanto menos inteligente deba ser el interprete mejor)

+++ src/pila/interprete/datos/Caracter.java
@@ -24,17 +24,17 @@ public class Caracter extends DatoPila {
 
     @Override
     public int toInt() {
-        return (int)valor;
+        return valor;
     }
 
     @Override
-    public long toNatural() {
-        return (long)valor;
+    public int toNatural() {
+        return valor;
     }
 
     @Override
     public float toFloat() {
-        return (float)valor;
+        return valor;
     }
 
     @Override
